Refactor code, optimize imports, change interfaces
socket's now return ssize_t like the underlying calls.

http_request no longer parsers a URL to figure out the host, protocol,
port, and resource. Now it just takes a cmd::stream which it assumes is
already connected. In the future I hope to add a function to do this
parsing but have it support other protocols like WebSocket (ws, wss).
The old code for parsing is now in resource_parser, but it is unused.

websocket now makes use of http_pool to get a connection and uses
http_request to send the upgrade request.

Move bind_server_socket to new file (server_socket.cc), previously in
tcp_server.

+++ examples/base64.cc
@@ -2,7 +2,6 @@
 #include <cstring>
 #include <iomanip>
 #include <iostream>
-#include <string>
 #include <vector>
 
 #include "base64.h"
@@ -13,7 +12,7 @@ int main(int argc, char *argv[])
 
     for (int i = 1; i < argc; i++) {
         auto t1 = high_resolution_clock::now();
-        int len = std::strlen(argv[i]);
+        size_t len = std::strlen(argv[i]);
         auto t2 = high_resolution_clock::now();
         std::string encoded11 = cmd::base64::encode(argv[i], len);
         auto t3 = high_resolution_clock::now();
